examples/research/requirements.txt
-e ../../
langgraph-cli[inmem]
tavily-python
langserve
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
